Enable no-extra-semi & no-irregular-whitespace
The no-extra-semi rule:
https://eslint.org/docs/rules/no-extra-semi
disallows unnecessary semicolons, e.g. it forbids the following:
```js
var x = 5;;

function foo() {
    // code
};
```

In jQuery we usually don't add a semicolon at the end of function declarations
but I found a few cases where we do, creating an inconsistency.

The no-irregular-whitespace rule:
https://eslint.org/docs/rules/no-irregular-whitespace
forbids most whitespaces that may be hard to notice in source. We have one such
occurrence in Core code that would be good to forbid.

Fixes #15